Both `nest` and `wrap` transformations add data to the existing subhash
1) `nest`

If the key, to which the data should be wrapped, is already exist
and contains a hash, then new data are added to that hash, while preserving
the existing data.

  fn = t(:nest, :foo, :baz)
  source = { foo: { bar: :BAR }, baz: :BAZ }

  # Old behaviour
  fn[source] # => { foo: { baz: :BAZ } }

  # New behaviour
  fn[source] # => { foo: { bar: :BAR, baz: :BAZ } }

In case the value under the key is not a hash, new data rewrite the old ones
as before:

  source = { foo: :FOO, baz: :BAZ }
  fn[source] # => { foo: { baz: :BAZ } }

2) `wrap`

The operation applies the updated `nest` to array of tuples.