Add `#name` attribute to the function
Because the function is defined locally from the module method, and
not necessary registered globally, it needs its own name to be carried
as an attribute.

The `Transproc::Registry.t` assings the name from its first argument:

  HashTransformations[:rename_keys, foo: :bar].name # => :rename_keys

in other cases the anonymous source (closure) is used as a function name.

  Transproc::Function.new(-> v { v }).name # => #<Proc ...>

There is also the possibility to set the name explicitly.

  Transproc::Function.new(-> v { v }, name: :identity) # => :identity