Prevent rename_keys from overwriting data
Fixes an issue where calling `rename_keys` could cause data in a hash to
be deleted.

`rename_keys` accepts a hash in which each key represents a key in the
value has that will be renamed *from*, and each value represents a key
that will be renamed *to*.

Data was deleted in the case that the "to" key already existed, but the
"from" key did not. This is because `delete(from)` was called on the
value hash; Ruby will return `nil` if `Hash#delete` is called with
a nonexistent key, and so `hash[to]` was always set to `nil` in these
cases.

`rename_keys` will now only set the "to" element in the hash if the
"from" key exists, preventing this issue from happening.